requirements.txt
Django==2.1.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
html5lib==1.0.1
Pillow==5.2.0
PyPDF2==1.26.0
pytz==2018.5
reportlab==3.5.6
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.3
